requirements.txt
phidata
langchain
langchain-community
groq
openai
langchain-groq
langchain-openai
python-dotenv 
docker
